DROID-3105 revert
