+++ domain/src/main/java/com/anytypeio/anytype/domain/primitives/FieldParser.kt
@@ -1,7 +1,5 @@
 package com.anytypeio.anytype.domain.primitives
 
-import com.anytypeio.anytype.core_models.Block
-import com.anytypeio.anytype.core_models.Id
 import com.anytypeio.anytype.core_models.MAX_SNIPPET_SIZE
 import com.anytypeio.anytype.core_models.ObjectType
 import com.anytypeio.anytype.core_models.ObjectWrapper
@@ -19,7 +17,6 @@ import kotlin.collections.contains
 interface FieldParser {
     fun toDate(any: Any?): Field.Date?
     fun getObjectName(objectWrapper: ObjectWrapper.Basic): String
-    fun getObjectName(map: Map<Id, Block.Fields>, objectId: Id?): String?
 }
 
 class FieldParserImpl @Inject constructor(
@@ -102,22 +99,6 @@ class FieldParserImpl @Inject constructor(
             return formattedDate
         } else {
             return ""
-        }
-    }
-    //endregion
-
-    //region Block.Fields
-    override fun getObjectName(map: Map<Id, Block.Fields>, objectId: Id?): String? {
-        return map.getProperObjectName(id = objectId)
-    }
-
-    private fun Map<Id, Block.Fields>.getProperObjectName(id: Id?): String? {
-        if (id == null) return null
-        val layoutCode = this[id]?.layout?.toInt()
-        return if (layoutCode == ObjectType.Layout.NOTE.code) {
-            this[id]?.snippet?.replace("\n", " ")?.take(MAX_SNIPPET_SIZE)
-        } else {
-            this[id]?.name
         }
     }
     //endregion
